[Refactor](Main): Simplified by separating on scenarios.
