+++ src/filetags/file_operations/__init__.py
@@ -1,4 +1,3 @@
-from .file_operations import *
 from .find_unique_alternative_to_file import *
 from .general import *
 from .links import *
