+++ src/filetags/scenarios/__init__.py
@@ -1,0 +1,3 @@
+from .common import *
+from .interactive import *
+from .process_files import *
