+++ src/filetags/scenarios/common.py
@@ -1,0 +1,6 @@
+# QUESTION: Change to pathlib?
+Files = list[str]
+# TODO: Make concrete.
+Tags = dict
+Vocabulary = dict
+TagsForShortcuts = list
